resources/code/workspace/puppy_guide/client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastmcp==3.0.2",
    "openai==2.21.0",
    "python-dotenv==1.2.1",
]
